Initialize PoolManager contract
