+++ contracts/evm/hardhat.config.ts
@@ -18,7 +18,7 @@ const mnemonic = process.env.LEDGER_SEED
 
 // Go to https://hardhat.org/config/ to learn more
 const config: HardhatUserConfig = {
-  solidity: '0.8.4',
+  solidity: '0.8.17',
   paths: {
     tests: './test',
     sources: './src',
