+++ contracts/evm/src/Sample.sol
@@ -1,5 +1,5 @@
-//SPDX-License-Identifier: Unlicense
-pragma solidity ^0.8.4;
+// SPDX-License-Identifier: Unlicense
+pragma solidity ^0.8.17;
 
 import "hardhat/console.sol";
