+++ contracts/evm/tsconfig.json
@@ -9,7 +9,7 @@
     "include": [
         "./deploy/*.ts",
         "./test/*.ts"
-    ],
+, "src/index.ts"    ],
     "typeRoots": [
         "node_modules/@types"
     ],
